+++ superset-frontend/src/views/CRUD/data/components/SyntaxHighlighterCopy/index.tsx
@@ -27,6 +27,7 @@ import github from 'react-syntax-highlighter/dist/cjs/styles/hljs/github';
 import SyntaxHighlighter from 'react-syntax-highlighter/dist/cjs/light';
 import { ToastProps } from 'src/messageToasts/enhancers/withToasts';
 import Icon from 'src/components/Icon';
+import copyTextToClipboard from 'src/utils/copy';
 
 SyntaxHighlighter.registerLanguage('sql', sqlSyntax);
 SyntaxHighlighter.registerLanguage('markdown', markdownSyntax);
@@ -61,41 +62,17 @@ export default function SyntaxHighlighterCopy({
   language: 'sql' | 'markdown' | 'html' | 'json';
 }) {
   function copyToClipboard(textToCopy: string) {
-    const selection: Selection | null = document.getSelection();
-    if (selection) {
-      selection.removeAllRanges();
-      const range = document.createRange();
-      const span = document.createElement('span');
-      span.textContent = textToCopy;
-      span.style.position = 'fixed';
-      span.style.top = '0';
-      span.style.clip = 'rect(0, 0, 0, 0)';
-      span.style.whiteSpace = 'pre';
-
-      document.body.appendChild(span);
-      range.selectNode(span);
-      selection.addRange(range);
-
-      try {
-        if (!document.execCommand('copy')) {
-          throw new Error(t('Not successful'));
-        }
-      } catch (err) {
+    copyTextToClipboard(textToCopy)
+      .then(() => {
         if (addDangerToast) {
           addDangerToast(t('Sorry, your browser does not support copying.'));
         }
-      }
-
-      document.body.removeChild(span);
-      if (selection.removeRange) {
-        selection.removeRange(range);
-      } else {
-        selection.removeAllRanges();
-      }
-      if (addSuccessToast) {
-        addSuccessToast(t('SQL Copied!'));
-      }
-    }
+      })
+      .catch(() => {
+        if (addSuccessToast) {
+          addSuccessToast(t('SQL Copied!'));
+        }
+      });
   }
   return (
     <SyntaxHighlighterWrapper>
